add error catch middleware.
